Made abstract class for resending notification, so implementation can be value stream dependend. Removed unneccesary pom dependency.
